Sound Changed+Minor Refactor+Improved Module Store
(the modules of your ship will now actually start in the store, but with
some bugs atm) like defaultModule not existing (haven't tested with
multiplayer)

+++ StellarReaction/src/Control/BasePlayerTraits.cpp
@@ -65,8 +65,8 @@ void BasePlayerTraits::addModule(const std::string& newTitle, const b2Vec2& rPos
 	m_owned.push_back(pair<string, b2Vec2>(newTitle, rPos));
 	sf::Packet pack;
 	pack << newTitle;
-	pack << (int32_t)rPos.x;
-	pack << (int32_t)rPos.y;
+	pack << (float)rPos.x;
+	pack << (float)rPos.y;
 	Message modAdded("ship_editor", "addItem", pack, 0.f, false);
 	game.getCoreIO().recieve(modAdded);
 }
